ACPI: Support PLIC in ACPI
Add a subtype in Madt to support PLIC. It needs to be clarified that this is not standard in acpi spec in which PLIC is not supported yet.
Add support for ACPI in irq-sifive-plic driver.
It should be noted that, currently, irq controller cascading is not supported because we can only get the default irqdomain fwnode in acpi_get_gsi_domain_id.
Now before PLIC was registered, the acpi_get_gsi_domain_id will return the fwnode of INTC, and once the PLIC is registered, the fwnode of PLIC will be returned.
In platform driver, linux irq number can be obtained through platform_get_irq.

Signed-off-by: Guo Yaxing <guoyaxing@bosc.ac.cn>