dts: add rngseed for early entropy init
The purpose of the rngseed device tree node:
It specifies the seed data source for the random number generator during early boot,
ensuring better entropy initialization for the system.

When running multi-core apache on PLDM, the accumulation of entropy on PLDM is very
slow, which causes geturandom to run for a long time. Add rngseed node to speed up
the initialization of geturandom function.

The relevant task links are as follows:
https://www.teambition.com/task/677b42884f91b245709c9517

Signed-off-by: YunFeng Yang <yangyunfeng@bosc.ac.cn>